refactor: rename `MoniteContext` to `MoniteAPIProvider`
